airflow-docker/requirements.txt
apache-airflow
python-dotenv
boto3
